Supoprt elasticsearch-keystore (#15)
* Support elasticsearch-keystore
fixes #9

* Fixed some issues regarding plugins and keystore (#16)

The plugin options property is already a string, so the `.join()` method
does not work. Removed it and added a new test case for this issue.

The `--f` flag on the `keystore add-file` is necessary because ES will
ask if we want to create/override the existing key-store. Since this
command will be executed in every deployment, we need to force it so
that ES is not expecting an input from the user.