Added authentication to user creation errand
If the ES cluster does not have an anonymous user with the proper write
permissions, we need to authenticate in order to perform the requests
that will create/update users.

So, an optional property was included here to allow authentication in
the curl calls.

By default, it's disabled. If it is enabled, then the username and
password are also needed.