Added errand job to create ES roles
This errand job will allow users to create 2 roles in ES:
one role that will be used by the anonymous user and another role that
will be used by the shared kibana user.

The roles are created with the permissions specified in the deployment.
The permissions should be added in JSON format, which will then be used
in the curl requests.

In the future this errand job should be changed to be able to iterate
over a list, giving more flexibility to the operator. But, for now, only
creates the 2 roles already mentioned.